## Description

> In import application flow we are using the save instead of saveAll even if we have the list of objects which can be saved at once. This PR updates the related execution to use saveAll in possible places.

Fixes #14363

## Type of change

- Performance optimization

## How Has This Been Tested?

> Tested the flow with application having 52 pages, following are the results for multiple execution:
Execution time before code changes:
6560ms, 6088ms, 7597ms, 6140ms
After code changes:
5701ms, 4375ms, 4487ms, 4912ms


## Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes